feat: instruction builders convenience function
