+++ packages/gill/src/index.ts
@@ -2,6 +2,7 @@ export * from "@solana/kit";
 
 export * from "./types";
 export * from "./core";
+export * from "./builders";
 
 // Node has a special export for the node JS specific
 // export * from "./node"; // DO NOT EXPORT HERE
